Add support for nested values + fix nullability of preview types
