+++ src/generate.ts
@@ -1,11 +1,12 @@
-import { AttributeTypes, Property, WidgetXml } from "./WidgetXml";
-import { extractProperties, toClientType } from "./xmlHelpers";
+import { generateClientTypes } from "./generateClientTypes";
+import { generatePreviewTypes } from "./generatePreviewTypes";
+import { extractProperties } from "./helpers";
+import { Property, WidgetXml } from "./WidgetXml";
 
 export function generateForWidget(widgetXml: WidgetXml, widgetName: string) {
     if (!widgetXml?.widget?.properties) {
         throw new Error("[XML] XML doesn't contains <properties> element");
     }
-
     if (widgetXml.widget.$.pluginWidget !== "true") {
         throw new Error("[XML] Attribute pluginWidget=true not found. Please review your XML");
     }
@@ -17,21 +18,16 @@ export function generateForWidget(widgetXml: WidgetXml, widgetName: string) {
         : []
     ).filter((prop) => prop && prop.$ && prop.$.key);
 
-    const enumTypes = Array.of<string>();
-    const clientTypes = properties
-        .map((prop) => {
-            const isOptional = (prop.$.required === "false" && prop.$.type !== "object") || prop.$.type === "action";
-            return `    ${prop.$.key}${isOptional ? "?" : ""}: ${generateClientType(prop, enumTypes, isNative)};`;
-        })
-        .join("\n");
-
-    const modelerTypes = properties
-        .map((prop) => {
-            const isAlwaysPresent =
-                prop.$.type === "enumeration" || (prop.$.type === "object" && prop.$.isList === "true");
-            return `    ${prop.$.key}${!isAlwaysPresent ? "?" : ""}: ${generateModelerType(prop)};`;
-        })
-        .join("\n");
+    const clientTypes = generateClientTypes(widgetName, properties, isNative);
+    const modelerTypes = generatePreviewTypes(widgetName, properties);
+    const visibiltyMap = `export interface VisibilityMap ${generateVisibilityMap(properties, "")}`;
+    const allTypes = isNative
+        ? clientTypes.join("\n\n")
+        : clientTypes
+              .slice(0, clientTypes.length - 1) // all client auxiliary types
+              .concat(modelerTypes.slice(0, modelerTypes.length - 1)) // all preview auxiliary types
+              .concat([clientTypes[clientTypes.length - 1], modelerTypes[modelerTypes.length - 1], visibiltyMap])
+              .join("\n\n");
 
     const mxImports = [
         "ActionValue",
@@ -44,143 +40,25 @@ export function generateForWidget(widgetXml: WidgetXml, widgetName: string) {
         "ObjectItem",
         "WebIcon",
         "WebImage",
-    ].filter((type) => new RegExp(`\\W${type}\\W`).test(clientTypes));
+    ].filter((type) => new RegExp(`\\W${type}\\W`).test(allTypes));
     const reactImports = Array.of<string>()
-        .concat(/\WComponent\W/.test(modelerTypes) ? ["Component"] : [])
+        .concat(/\WComponent\W/.test(allTypes) ? ["Component"] : [])
         .concat(!isNative ? ["CSSProperties"] : [])
-        .concat(/\WReactNode\W/.test(clientTypes) ? ["ReactNode"] : []);
+        .concat(/\WReactNode\W/.test(allTypes) ? ["ReactNode"] : []);
     const imports = [
         `import { ${reactImports.join(", ")} } from "react";`,
         `import { ${mxImports.join(", ")} } from "mendix";`,
-    ].filter((line) => line.indexOf("{  }") === -1);
+    ]
+        .filter((line) => line.indexOf("{  }") === -1)
+        .join("\n");
 
-    if (isNative) {
-        return `/**
- * This file was generated from ${widgetName}.xml
- * WARNING: All changes made to this file will be overwritten
- * @author Mendix Content Team
- */
-${imports.length ? imports.join("\n") + "\n\n" : ""}${enumTypes.length ? enumTypes.join("\n\n") + "\n\n" : ""}
-export interface ${widgetName}Props<Style> {
-    name: string;
-    style: Style[];
-${clientTypes}
-}
-`;
-    }
     return `/**
  * This file was generated from ${widgetName}.xml
  * WARNING: All changes made to this file will be overwritten
  * @author Mendix Content Team
  */
-${imports.length ? imports.join("\n") + "\n\n" : ""}${enumTypes.length ? enumTypes.join("\n\n") + "\n\n" : ""}
-export interface ${widgetName}ContainerProps {
-    name: string;
-    class: string;
-    style?: CSSProperties;
-    tabIndex: number;
-${clientTypes}
-}
-
-export interface ${widgetName}PreviewProps {
-    class?: string;
-    style?: string;
-${modelerTypes}
-}
-
-export interface VisibilityMap ${generateVisibilityMap(properties, "")}
+${imports.length ? imports + "\n\n" : ""}${allTypes}
 `;
-}
-
-function generateClientType(prop: Property, enumTypes: string[], isNative: boolean): string {
-    switch (prop.$.type) {
-        case "boolean":
-            return "boolean";
-        case "string":
-            return "string";
-        case "action":
-            return "ActionValue";
-        case "textTemplate":
-            return "DynamicValue<string>";
-        case "integer":
-        case "decimal":
-            return "BigJs.Big";
-        case "icon":
-            return isNative ? "DynamicValue<NativeIcon>" : "DynamicValue<WebIcon>";
-        case "image":
-            return isNative ? "DynamicValue<NativeImage>" : "DynamicValue<WebImage>";
-        case "file":
-            return "DynamicValue<FileValue>";
-        case "datasource":
-            return "ListValue";
-        case "attribute":
-            if (!prop.attributeTypes || prop.attributeTypes.length === 0) {
-                throw new Error("[XML] Attribute property requires attributeTypes element");
-            }
-            return `EditableValue<${toClientTypes(prop.attributeTypes[0])}>`;
-        case "expression":
-            if (!prop.returnType || prop.returnType.length === 0) {
-                throw new Error("[XML] Expression property requires returnType element");
-            }
-            return `DynamicValue<${toClientType(prop.returnType[0].$.type)}>`;
-        case "enumeration":
-            generateEnums(prop, enumTypes);
-            return capitalizeFirstLetter(prop.$.key) + "Enum";
-        case "object":
-            const childType = capitalizeFirstLetter(prop.$.key) + "Type";
-            // todo
-            return prop.$.isList === "true" ? `${childType}[]` : childType;
-        case "widgets":
-            return !!prop.$.dataSource ? "(item: ObjectItem) => ReactNode" : "ReactNode";
-        default:
-            return "any";
-    }
-}
-
-function generateModelerType(prop: Property): string {
-    switch (prop.$.type) {
-        case "boolean":
-            return "boolean";
-        case "string":
-            return "string";
-        case "action":
-            return "{}";
-        case "textTemplate":
-            return "string";
-        case "integer":
-        case "decimal":
-            return "number";
-        case "icon":
-            return "IconProperty";
-        case "image":
-            return "ImageProperty";
-        case "file":
-            return "string";
-        case "datasource":
-            return "ListValue"; // todo
-        case "attribute":
-        case "expression":
-            return "string";
-        case "enumeration":
-            return capitalizeFirstLetter(prop.$.key) + "Enum";
-        case "object":
-            const childType = capitalizeFirstLetter(prop.$.key) + "PreviewType";
-            // todo
-            return prop.$.isList === "true" ? `${childType}[]` : childType;
-        case "widgets":
-            return "({ widgetCount: number; renderer: Component<{}> })";
-        default:
-            return "any";
-    }
-}
-
-function generateEnums(prop: Property, childTypes: string[]) {
-    if (!prop.enumerationValues?.length || !prop.enumerationValues[0].enumerationValue?.length) {
-        throw new Error("[XML] Enumeration property requires enumerations element");
-    }
-    const typeName = capitalizeFirstLetter(prop.$.key) + "Enum";
-    const members = prop.enumerationValues[0].enumerationValue.map((type) => `"${type.$.key}"`);
-    childTypes.push(`export type ${typeName} = ${members.join(" | ")};`);
 }
 
 function generateVisibilityMap(properties: Property[], indent: string): string {
@@ -200,18 +78,4 @@ function generateVisibilityMap(properties: Property[], indent: string): string {
             .join("\n") +
         `\n${indent}}`
     );
-}
-
-function capitalizeFirstLetter(text: string): string {
-    return text.charAt(0).toUpperCase() + text.slice(1);
-}
-
-function toClientTypes(attributeTypes: AttributeTypes): string {
-    if (!attributeTypes.attributeType?.length) {
-        throw new Error("[XML] Attribute property requires attributeTypes element");
-    }
-    const types = attributeTypes.attributeType
-        .filter((type) => type.$ && type.$.name)
-        .map((type) => toClientType(type.$.name));
-    return Array.from(new Set(types)).join(` | `);
 }
